split config on testnet,mainnet,default
