+++ config/mainnet/config.js
